requirements.txt
# Core
numpy>=1.26
pyyaml>=6.0.1
pandas>=2.2
scikit-learn>=1.5

# Deep learning
torch>=2.3,<2.4

# Simulation
pybullet>=3.2.5

# Visualization & notebooks
matplotlib>=3.8
seaborn>=0.13
plotly>=5.22
jupyter>=1.0

# Experimentation & logging
tensorboard>=2.16
tqdm>=4.66
rich>=13.7
h5py>=3.10

# CLI & tooling
typer>=0.12
pytest>=8.2
